// types/AccountSqlRecord.ts
import SqlRecord from '@/types/SqlRecord'
export default interface AccountSqlRecord extends SqlRecord {
  moved_to_account_id?: string
  uri: string
  url?: string
  username: string
  secret: string
  public_key: string
  remote_url: string
  salmon_url: string
  hub_url: string
  note: string
  display_name: string
  locked: boolean
  header_remote_url: string
  inbox_url: string
  domain?: string
  private_key?: string
  avatar_file_name?: string
  avatar_content_type?: string
  avatar_file_size?: number
  avatar_updated_at?: string
  header_file_name?: string
  header_content_type?: string
  header_file_size?: number
  header_updated_at?: string
  avatar_remote_url?: string
  subscription_expires_at?: string
  last_webfingered_at?: string
  outbox_url: string
  shared_inbox_url: string
  followers_url: string
  protocol: number
  memorial: boolean
  featured_collections_url?: string
  fields?: string
  actor_type?: string
  discoverable?: boolean
  also_known_as?: string[]
  silenced_at?: string
  suspended_at?: string
  trust_level?: number
}